Update 'Chef Server' to 'Chef Infra Server' in the docs
All instances of 'Chef Server' are updated to 'Chef Infra Server' except
where they refer to UI elements in the software which wouldn't have been
changed.

This is to align with our use of 'Chef Infra' to describe the 'Chef'
project to provide more specific clarity to users.

A few but not all 'Chef Client' to 'Chef Infra Client' changes are also
made in this commit.

Signed-off-by: Bryan McLellan <btm@loftninjas.org>