Check-in time series total nodes (#3179)
Adding the total number of nodes for each daily range. To calculate the number of nodes for each range find the number of nodes that have been deleted and created before the end of the range. Subtract the number of nodes created from the number of nodes delete and you have the total number of nodes. To find the number of nodes created, a create date was added to the node-state. During ingestion, if the node is new it adds the create date of the time now. If the node already exists it does not update the create date.
```
{
  "counts": [
    {
      "start": "2020-03-20T19:00:00Z",
      "end": "2020-03-21T18:59:59Z",
      "count": 0,
      "total": 0 <-------
    },
    {
      "start": "2020-03-21T19:00:00Z",
      "end": "2020-03-22T18:59:59Z",
      "count": 5,
      "total": 5 <-------
    },
    {
      "start": "2020-03-22T19:00:00Z",
      "end": "2020-03-23T18:59:59Z",
      "count": 6,
      "total": 6  <-------
    }
  ]
}
```

Note: Copied one of olivere's go files to add the additional "missing" option to allow for counting nodes that do not have a create date. https://github.com/olivere/elastic/blob/release-branch.v6/search_aggs_bucket_date_range.go

Below is a link the the "missing" option information.
https://www.elastic.co/guide/en/elasticsearch/reference/7.0/search-aggregations-bucket-daterange-aggregation.html#_missing_values_3

Signed-off-by: Lance Finfrock <lfinfrock@chef.io>