studio: generate gateway mocks using Makefile
In d1a06544e5f009956d4a7b2675f85fc4fa5cb882 we added a mockgen build
tag to avoid various chicken-and-egg problems that are the result of
the current package structure in the gateway. The mockgen tag allows
us to skip various files not required for generating the mocks but
which create problems during the build.

Rather than re-add that same logic to the studio function, we now call
out to the Makefile.

I've removed two make targets that no longer appear to be used since
the relevant protobufs have moved.

Signed-off-by: Steven Danna <steve@chef.io>