Dan/top infra errors (#3195)
API for the top Chef Infra errors. Aggregates the most common error by class+message among the most recent chef runs across the fleet. The nodes considered in the aggregation can be filtered with the same filters as other node APIs (except status is not allowed because the aggregation only considers nodes with "failed" status).

Commits:

* Add stub API for top config mgmt errors
* Store CCR error class separately in node-state index
* Implement top errors API
* Implement filtering for top errors API
* implement size option for get errors API
* Update bldr config