Fix builder backup bug (#3250)
The backups were failing trying to write the snapshot files. The issue
was that there was a bug where the blob writer was returning the number
of bytes written as greater than the buffer we passed in, thus confusing
the MultiWriter object. This seems to be fixed in the latest go-cloud
library.

Fixes chef/customer-bugs#154

Signed-off-by: Jay Mundrawala <jmundrawala@chef.io>