Modify builder oauth flow to exchange code in /token
Previously, we were exchanging the code received from dex in the
/callback endpoint. This exchange gave us a token that would need to be
cached, as it was supposed to be returned in the /token endpoint. This
caching meant that this flow wouldn't work in HA as the request could
get routed to any node, and each node was caching locally in memory.

Now, the call to /callback simply redirects the user to /bldr with the
unexchanged code. This code is given to bldr which calls /token which
does the token exchange with dex.

Signed-off-by: Jay Mundrawala <jmundrawala@chef.io>