studio: inline chef/studio-common code that we actually use
It appears that we only about 400 of the 1800 lines of shell scripts
in studio common. And, my guess is we use even less than that since
the functions we require handle many cases that we never actually
need.

In order to make it easier to debug the studio and have a complete
view of the code involved, this commit removes the relevant functions
by inlining the relevant functions.

Signed-off-by: Steven Danna <steve@chef.io>