Desktop UI Rough Insight Control (#3315)
Adding the desktop insight control. This control displays a filtered list of desktops. The control is initially hidden and appears when a row is selected from the "Top Errors" control. The desktops can be sorted by clicking one of the three column headers. This can be sorted in ascending and descending order. Pagination is also added. There is also a button on the top right to re-hide the control.

Signed-off-by: Lance Finfrock <lfinfrock@chef.io>