[es-sidecar] only verify snapshot repositories on external clusters
Previously we were verifying snapshot repositories twice on every call
to create a snapshot. Snapshot verification is only useful on multi-node
clusters because it ensures that all master/data nodes in the cluster
are functional. If there's an issue with the snapshot repository on a
single node cluster the snapshot will immediately fail when the request
is given so there's no need to verify it before requesting the snapshot.

* Make snapshot repository verification configurable at the es-sidecar
  level
* Don't verify snapshot repositories twice when creating a snapshot
* Automatically enable snapshot repository verification for external
  clusters

Signed-off-by: Ryan Cragun <ryan@chef.io>