Treat UI unit test warnings as errors
With unit test warnings not causing build failures,
it is easy to either postpone fixing them indefinitely
or possibly not even notice them in the first place.
This somewhat complex machination adjusts the unit test
run so that not only errors but warnings
will now also return a non-zero exit code.

Signed-off-by: michael sorens <msorens@chef.io>