[backup] Add initial artifact repository integrity checking (#3276)
The shared artifact snapshot repository is used for builder artifact
backup operations. The current snapshot implementation determines
which artifacts _should_ exist in the shared repository by aggregating
the artifacts needed for all existing snapshots and doing a diff on the
existing assumed artifacts list with the required list of artifacts in the
new snapshot. Any artifacts that are not assumed present in the backup
repository are then uploaded from builder. The speed of this check is
desirable as it doesn't require listing every artifact in the shared repository
for each backup, an operation that will gradually become more expensive,
especially if the snapshot repository is in S3. The downside of this method
is that we make an assumption that the shared repository has all of
the artifacts that make up the aggregate list.

This change introduces basic shared repository integrity validation
checking. It scans the backup artifact repository and validates any
and/or all snapshots with the actual repository artifacts and saves
the latest validation check results in the snapshot bucket.

`chef-automate backup integrity validate` will validate all snapshots in
the repository. A user may also choose to exclusively validate one or
more snapshots by passing backup IDs as variable arguments to the
command.

`chef-automate backup integrity show` will pretty print the current
snapshot validation metadata. Eventually it might be nice to allow this
command to take backup IDs and to show all the details of a snapshot,
including which artifacts are missing if it is corrupted.

In the future we should look to add additional commands like `repair`
which will validate the snapshot and attempt to upload any missing
artifacts from the existing builder bucket.

This change also includes a new builder scaffolding tool called
`generate real` which is a one stop shop for priming a development
builder instance with real habitat origins, keys and packages.

`go run ./tools/builder-scaffolding/... generate real ./tools/builder-scaffolding/habitat/ `
will automatically build the shim package `builder-scaffolding` and
upload it to builder. It takes care of hab key, tokens, login, origins etc.

To facilitate a few functions in the new generation tool
`fileutils.CopyFile()` and `fileutils.CopyDir()` have been added.

The `ReplayableStream` stream type has been added which allows the same
stream to be repeatably read over and over. It's syntactic sugar over
existing streams so we can cache long I/O stream reads and replay it
locally many times. It was used to handle reading the remote bucket
objects once and replaying it for each snapshot that is validated.

Signed-off-by: Ryan Cragun <ryan@chef.io>