Config mgmt Get Node Metadata Counts (#3360)
Adding a new endpoint that for a node's field type provides distinct values and counts.

For each type of field requested this returns distinct values the amount of each. For example, 
if the 'platform' field is requested 'windows' 10, 'redhat' 5, and 'ubuntu' 8 could be returned. 
The number next to each represents the number of nodes with that type of platform.


Signed-off-by: Lance Finfrock <lfinfrock@chef.io>