Revisit and streamline
Thanks to question from @srenatus in the PR,
I got to thinking that there was more here than was needed.
Still need just one input parameter rather than two,
but switched to the other one, which is more indicative
of what it is doing:
this is a modal for projects and the checkbox is just implementation.
And when the modal opens--a state we already recognized,
we just need to set the checkbox true; no additional subscritption needed.

Signed-off-by: michael sorens <msorens@chef.io>