pretty-proto: workaround relative import confusion in notifications
The relative imports used in notifications service means that we need
to add the notifications service proto dir to the import
path. However, this then causes conflicts as both the imported version
of the proto and the version we have explicitly passed on the command
line define the same types (since they are the same file).

Signed-off-by: Steven Danna <steve@chef.io>