UX: add policy id to differentiate "add" messages
If you specify multiple policies to add the new token to,
you would see a series of messages all saying just "Added new token",
one for each policy, which looks strange.

But even in the likely more common case of adding to a single policy,
it still looks add. Added new token... to what?
So now it says "... to 'x' policy".

Signed-off-by: michael sorens <msorens@chef.io>