automate-ui/teams/add-users: feed payload into success action
Assuming that the UI knows the current state of team membership, this
change will fix the wrong notification:

For a team with n users, having added another k users, the notification
would have read

    "Added k+n users to team."

Now, we'll just say

    "Added k users to team."

Note that this may or may not be true: if another actor (using the UI or
the API) had changed team memberships at the same time, it might now be
accurate. But since this problem always exists, I think it's also
acceptible here.

Signed-off-by: Stephan Renatus <srenatus@chef.io>