Resolve redux devtools crashing
Before crashing it sometimes complains it is going to be slow
and consume a huge amount of memory.
Needed to isolate the actions and state generating large data footprints
and prune those using action/state sanitizers.
That is described in the troubleshooting section of the docs.
As it took me some time to work through adapting the supplied info to
Automate, I pushed an update to their docs as well.

https://github.com/zalmoxisus/redux-devtools-extension/blob/master/docs/Troubleshooting.md#excessive-use-of-memory-and-cpu

Signed-off-by: michael sorens <msorens@chef.io>