[automate-cs-nginx] Pin shared habitat dependencies
This PR pins All dependencies shared between this package and the
chef-server-* packages to the versions required by the chef-server-*
packages.

Currently, the chef-server-* packages we are pinned to were hand-built
and any update to any of these three dependencies will require them to
be re-built.

We cannot follow the latest chef-server-* packages until the
following issue is resolved:

    https://github.com/chef/automate/pull/2002#issuecomment-600343881

Signed-off-by: Steven Danna <steve@chef.io>