Tweak navbar and sidebar styles to handle smaller viewports (#3654)
Until a more permanent solution is implemented, this commit retrofits the main navbar and sidebar styles to better handle smaller viewport sizes. The bulk of this diff is tweaks to paddings and font-sizes within Carbon grid breakpoints. The largest visual change is the main navbar collapsing to two rows within smaller viewports.

Signed-off-by: Scott Christopherson <scott@chef.io>