Updates to desktop dashboard grid layout
This commit contains an incremental update to the layout skeleton of the Desktop dashboard view and to the components displayed within it. It utilizes the Carbon grid system and some generic components already available within Automate to bring the functionality and styling of this view a little bit closer to the final production implementation.

Signed-off-by: Scott Christopherson <scott@chef.io>