Convert active state from binary to trinary
Finally, this lets us have multiple dropdown components coexisting.
We need two things:
1. Trigger the clickOutside handler on all dropdowns.
Removing the event propagation guard in the last commit took care of this.
2. Recognize the one we are actually opening and allow it.
That is handled by this commit.
The toggleDropdown fires first, and that signals that we are opening.
Then handleClickOutside fires, but since we have not yet fully opened,
it does not try to close us!
Once we have gone past that check, we can then indicate we are open.

Signed-off-by: michael sorens <msorens@chef.io>