enable project filtering on nodemanager nodes
a few weeks ago we introduced the ability to use project filtering
on manually added nodes by supporting the environment tag on those nodes
and appropriately applying it on the node reports.
what we did not do at that time is also apply the project filtering
on the nodemanager nodes, which affects the list present at scan jobs/nodes
and integration/nodes.
this change ensures that we apply the project filtering to the
nodemanager nodes as well.

Signed-off-by: Victoria Jeffrey <vjeffrey@chef.io>