return count of nodes even if nodes result is 0
when the compliance reporting nodes api is called with a status
that has 0 nodes associated with it, the api was returning 0 for all the counts.
the counts returned should not be affected by status filters, only the list should.
this change modifies the code to ensure that we fetch the correct node counts regardless
of how many nodes were found to match the status in es

Signed-off-by: Victoria Jeffrey <vjeffrey@chef.io>