Desktop Compliance Profiles Page Update (#3809)
Update the Compliance Profile page to include the case of the Desktop Offering is installed. This also updates wording and style for other cases. We want to hide and show the search bar, "X Profiles" "X Available" tabs, profile table, upload button, and text depending on if the Desktop Offering is installed, one or more profiles are installed, and the user has permissions to upload a profile.

Signed-off-by: Lance Finfrock <lfinfrock@chef.io>