Notifications service compliance update (#3881)
InSpec reports have been updated to have truncated results. This PR is updating the number of control tests based on truncated results added structure from this PR #3595.

A bug was found and fixed with this work. The number of critical failed control tests was not counted correctly. The code was instead counting the number of critical failed controls.

Signed-off-by: Lance Finfrock <lfinfrock@chef.io>