Apps: use wrapper for new bool config setting
Users may have existing code to POST a body like the following to
configure disconnected services settings:
```
{
  "threshold": "15m"
}
```
This would result in any new plain bool value being set to false. Using
wrappers lets the application determine whether the value was set or
not, so we can maintain the old behavior.

Signed-off-by: Daniel DeLeo <dan@chef.io>