Include Services Data in Platform-wide Data Lifecycle API (#3931)
Integrate the applications-service's data lifecycle settings APIs with the new Automate-wide data lifecycle APIs. With this in place we can remove the grey-out on the configuration section for these options in the settings page.

* Update applications periodic job API ddls to match data lifecycle semantics
* apps tab - data lifecycle - return job info w/ config GET
* Apps: return recurrence of periodic jobs in GET APIs
* Apps: use wrapper for new bool config setting

Users may have existing code to POST a body like the following to
configure disconnected services settings:
```
{
  "threshold": "15m"
}
```
This would result in any new plain bool value being set to false. Using
wrappers lets the application determine whether the value was set or
not, so we can maintain the old behavior.

* Apps tab: make periodic job recurrence settable
* Validate config updates for applications periodic jobs
* Make running-ness of disconnected services job settable
* Update apps tab data lifecycle integration tests for new capabilities
* Add apps client to data lifecycle gateway server
* Implement "run job" APIs for apps service periodic tasks
* Add applications/services to std data lifecycle facade API
* gofmt -s
* Fix test configuration of apps periodic job runner
* Update embedded docs in apps protos
* Update data lifecycle docs with services lifecycle APIs