send reports to notifications after we have glued them back together
notifications service was receiving reports with values missing, such
as impact. we realized hte reason for this is that when we added logic to allow
for smaller, more compact reports to be ingested, we forgot to move the call
to notifications client further down the process to after we glue all
the information back together again.

Signed-off-by: Victoria Jeffrey <vjeffrey@chef.io>