add endtime string to compliance report
this is a thing that has bothered me for a while.
it recently came up again when
we needed to look at some info for the customer.
we translate the end time to grpc timestamp before sending it out,
making it return the epoch seconds for the end time value on the report.
having a reference to the epoch seconds of when a report was created
is not super useful to the customer, so this adds the end time as a string.
did not remove the original end time fields bc backwards compat

Signed-off-by: Victoria Jeffrey <vjeffrey@chef.io>