Fix ontop backups for builder (#4061)
If you were to take a backup of builder, then on another machine, install automate + builder, and restore the backup, it would fail because we'd change the minio secret. Minio doesn't like this because it encrypts config with that secret now. We don't actually care about that config, so this PR detects that case and blows away the config