Fix Invalid credential in backup gateway
minio uses the access and secret key to encrypt its config. We do
not need this. We work around this by storing those values in our
tmp directory that gets put in the folder.
Then, when we start up the backup gateway, we can check for that
file. If it exists, we tell backup-gateway that it should use those
values.
If we were not to do this, then it would be impossible to have 2
automates looking at a shared network drive.

This is a best effort thing. If the Invalid credential persists,
the best thing that can be done is to turn automate off, remove the
.tmp directory in the backups, and turn automate back on.

Signed-off-by: Jay Mundrawala <jmundrawala@chef.io>