Wrap os/user to fall back to getent (#4110)
Partial fix for #4105

If the hab user and groups and not local to the system, automate is unable to start. A good portion of the problem is that we use golang's `os/user` library to get UIDs and GIDs, which just looks at `/etc/passwd` and `/etc/group` when compiled without CGO. The bulk of this PR works around that. Instead of using `os/user`, we create out own user library which first tries `os/user`, the falls back to shelling out to `getent`. We can't shell out to any getent though, we need to find a system one as its going to know how to deal with the systems `/etc/nsswitch.conf`.

Unfortunately, this does not fix everything. The habitat supervisor links against a glibc that isn't going to lookup the nss modules (`libnss_*`) on the system. That means just because your system tools say a user/group exist, doesn't mean that hab will agree. To get around this, for now, we need to lay down a systemd unit overlay:
```
cat > /etc/systemd/system/chef-automate.service.d/custom.conf <<EOF
[Service]
Environment=LD_PRELOAD=/lib64/libnss_ldap.so
EOF
```
Doing this also makes all the places we depend on `core/coreutils` work. I think our main use-case would be `chown`.
Alas, we use busybox in a bunch of places to get chown, the `LD_PRELOAD` wont help.

Things that will happen in another PR:
- Stop using busybox and perhaps create a `chown` that can benefit from our `getent` hack in golang. Its also unclear we need to be using chown in a lot of places that we use it. automate-elasticsearch and automate-es-gateway currently complain, but automate-postgresql, automate-deployment, automate-cs-nginx, automate-ui, and automate-build-api-proxy need to be audited.
- If nscd is running on the box, coreutils will do the right thing