ci: move upload-files.sh to a buildkite pipeline
Our upload-files.sh script is suddenly 6+ times slower when run via a
bash action than it was previously. We think this is because of slower
disk speeds in the new expeditor setup.

Moving this script to buildkite gets around this limitation and comes
with a few other benefits:

- streaming logs during the script execution
- ability to easily upload debugging data as buildkite artifacts
- ability to easily parallelize the process if it becomes slower

We now trigger the post-promote pipeline after the
promote_habitat_packages pipeline. The fastly cache purge, deploy/dev,
and deploy/automate-chef-io pipelines then trigger on the successful
completion of the post-promote pipeline.

I've moved the code coverage step into this pipeline just to cut down
on the number of pipelines we need to look at in buildkite. Also, it
didn't appear to me that the code-coverage job needed to be in the
same concurrency group as it shouldn't affect the shared deploy
infrastructure at all so I moved it to its own concurrency group.

Signed-off-by: Steven Danna <steve@chef.io>