Test for Idempotent Schema Migration (#4187)
In https://github.com/chef/automate/pull/4183 we fixed an error in the migrator that was re-running migrations in multiple front-end cluster topologies, but after that patch we still try to forcibly redo migrations if they fail and leave a dirty schema state. So it's best if our migrations are able to re-run. This PR adds a function to our migrator library which drops all the data in the database, then runs a service's migrations all the way through twice. In addition to checking that migrations are idempotent, this approach makes it convenient to develop new migrations, you can keep re-running the integration tests and any bad state created by mistakes gets cleaned up.

I only modified the two services I know well in this pass, it should be pretty easy to add to others if we like.

* Add test for migration idempotence to the migrator lib
* Make config-mgmt's migrations idempotent and test it
* Add idempotent migration checking to applications test suite