chef-server: Upgrade to latest current build (#4216)
* chef-server: Upgrade to latest unstable build

This upgrades us to the latest unstable packages built from master.
To do so, it:

- Adds server_name_indication SSL configuration to avoid the issue
  outlined in chef/automate#2002.

- Adapts our various wrapper scripts to changes in the upstream
  packages around ruby+bundler. Chef Server is currently in the
  process of trying to upgrade many of its deps, so more churn here is
  likely.

- Sets the new retry_on_conn_closed option for oc_httpc to true, which
  hopefully works around some recent ibrowse changes.

Signed-off-by: Steven Danna <steve@chef.io>

* Add core/ruby26 exception

We already allow core/ruby, this is just a version upgrade where the
package name changed.

Signed-off-by: Steven Danna <steve@chef.io>

* chef-server-ctl: fix argument processing

This code was trying to run `knife opc "org user" add. Now, we split
on whitespace to make sure we pass "org user" as seperate args. I
believe this was working before because of subtle differences in the
wrapper scripts.

Signed-off-by: Steven Danna <steve@chef.io>

* automate-chef-server: Use new sqerl verify_ca feature

This allows us to avoid a breaking changing by continuing to verify
certificates but ignoring hostname mismatches.

Signed-off-by: Steven Danna <steve@chef.io>

* erchef,bifrost,bookshelf: configure cipher list

Signed-off-by: Steven Danna <steve@chef.io>

* add missing ,

Signed-off-by: Steven Danna <steve@chef.io>

* fix umask before install packages

If we don't set the umask, directories in /hab/pkgs/ might become
unreadable by the hab user.

Signed-off-by: Steven Danna <steve@chef.io>

* erchef,bifrost,bookshelf: configure eccs

This limits the curves offered via TLS in accordance with our TLS
scanner.

Signed-off-by: Steven Danna <steve@chef.io>