ci: fix ha cluster test failures
The argument parsing in Habitat has changed such that the command we
were using to create the database user actually displayed the
help (because of our use of the `-h` flag intended to be passed to
`psql` but now parsed by hab).  The help output returns an exit code
of 0, meaning that the user went uncreated without an error being
returned by our setup script.

Signed-off-by: Steven Danna <steve@chef.io>