Resurrect Event Feed Download PR from being reverted (#4237)
This PR was reverted because the PR it was based on had to be reverted. The original PR is #4181. The PR was based on is #4108.

Adds a new endpoint that allows CSV and JSON download of the events.
Moves the event feed protobufs from the automate gateway to the api/external directory
Adds the chef org and chef server URL to the event feed API.

Signed-off-by: Lance Finfrock <lfinfrock@chef.io>