Bump chef-server versions
This pulls in a workaround for TLS issues with external automate. The
workaround requires some new configuration with restores the old
behavior where we validate the certificate against the configured CA
but ignore hostname mismatches.

Our hope is to have full hostname checking support when Chef Server
upgrades to Erlang 22.

Signed-off-by: Steven Danna <steve@chef.io>