Removing inner transaction
When the transaction failed a new one was created inside the existing one.
This setup was found to be a problem. The solution was to create a function
that first tries to insert the message with the ID. If there is no row with
the ID an insert is tried with the cloud information. If the cloud information
is a duplicate that duplicate row is update.

Signed-off-by: Lance Finfrock <lfinfrock@chef.io>