Removing inner transaction (#4289)
What we are trying to solve with this PR is removing the piece of code that creates a transaction inside of a transaction.
When the outer transaction failed a new one was created inside the existing one. The solution was to create a postgresql function that first tries to insert the message with the ID. If there was no row with the ID, insert the row with the cloud information. If there is already cloud information matching, that matching row is updated.

Signed-off-by: Lance Finfrock <lfinfrock@chef.io>