Stop Vendoring Go Dependencies (#4314)
We want to be able to split the Automate repo into multiple golang modules. Given the current capabilities/behaviors of the go module tooling, having multiple modules in the same repo is a bad experience when using vendoring, because the local code is also vendored and must be manually updated. We generally prefer to vendor when we can but at this time it's holding us back. If the module system gains improved behaviors for working with local modules and vendoring, then we should reconsider.

Commits:
* Nuke the vendor dir
* Fix proto build and bldr config for novendor
* Remove references to the deleted go vendor dir
* Appease license scout
* Set proxy vars before using any go install in tests