Make external API a separate module (#4298)
Make the api/external directory a separate golang module. When using golang modules, when you depend on any package within a module, you depend on that module and therefore also all of that module's dependencies. Currently all of Automate is one big module which is nice for our internal dependency management but for external users it means you have to take on a lot of deps even if you only intend to use a few structs from api/external directory.

At present the goal is to support a project that is integrating some Chef Workstation functionality with Chef Automate, but this would be helpful to anyone who wishes to write code using Automate's API in go.

* Make external API a separate module
* Revendor deps after extracting API module
* Update licensing for api/external module
* license scout hints for chzyer/logex
* Moar license scouting