Fixing Integration test surprise (#4424)
Two integration tests in the config-mgmt-service started failing on November 3rd because of the residual of other tests. The other tests were creating Elasticsearch indexes for a static date of November 3rd, 2019, but the tests that just started to fail were assuming there were no indexes for the past year from now. The fix was to set the failing tests to a static date.

Signed-off-by: Lance Finfrock <lfinfrock@chef.io>