Apply default TLS cipher suite to nats-gateway
This got picked up in a recent security scan. Someone familiar with
the application should review to ensure there wasn't a reason to allow
the less restrictive cipher suite.

Note that this still leaves the NATS ports 4222 and 10140 outside the
view of the scanner as the scanner says those don't appear to be TLS
ports.

Signed-off-by: Steven Danna <steve@chef.io>