Jamie/4366 disappearing pagination (#4441)
* Add a check against last page to show pagination component

There is an issue where if there is only one selectable page, the pagination component is hidden.  As a result if there are say...11 pages total, and the user is shown a max of 10 selectable pages at time, when they get to page 11, the component is only showing 1 selectable page, and the component hides.

Signed-off-by: seajamied <jdegnan@chef.io>

* Update page number to 1 when a new filter is selected

I cant think of an apporpriate way to calculate the current page with the given information.  This solves the problem, but i'm not sure how i feel about it yet.  I think there may be a better way and Id love to remain on the current page if possible

Signed-off-by: seajamied <jdegnan@chef.io>

* Add an additional 'current page' calculation

This fixes the UX portion, so that a user doesn't have to get thrown back to the 1st page for this to work
when a new selection is made. this should not affect other implementations of the page picker component.

Signed-off-by: seajamied <jdegnan@chef.io>

* Update variable name to maxSelectablePages for better clarity

Signed-off-by: seajamied <jdegnan@chef.io>

* Change calc to use max min

Signed-off-by: seajamied <jdegnan@chef.io>

* Do not disable a page selector if there are multiple selectable pages

Signed-off-by: seajamied <jdegnan@chef.io>

* Clean up the calculations a bit

Signed-off-by: michael sorens <msorens@chef.io>

Co-authored-by: Michael Sorens <msorens@chef.io>