Studio function to create an example Policyfile (#4465)
Adding a new studio functions push_example_policyfile and converge_policyfile_chef_client that create a CCR with an example policyfile. This is useful when testing policyfiles with Infra nodes or the event feed. The converge_policyfile_chef_client function first calls the push_example_policyfile function to create the example policyfile and upload it to Chef-Server.

Signed-off-by: Lance Finfrock <lfinfrock@chef.io>