deployment-service/ci: retry file-access errors when waiting for upgrade
During upgrades from old versions, we may require a deployment-service
certificate regeneration. If the chef-automate command line tool
attempts to read the deployment-service certificate during this
regeneration, it may receive a FileAccessError such as:

    FileAccessError: Unable to access the file or directory: Failed to
    read deployment-service TLS certificates: Could not read the
    service cert: open
    /hab/svc/deployment-service/data/deployment-service.crt: no such
    file or directory

Since we expect this error to be transient, we now retry this error in
the wait_for_upgrade loop in CI. We also log a bit more information
about the deployment-service certificate regen.

Signed-off-by: Steven Danna <steve@chef.io>