Remove rebase-on-the-fly in PRs
While this was useful to have during periods of heavy development,
the PR traffic is now much reduced so it is OK
to change this back to the way almost every other Chef repo works.

The motivation for this was for tools (like semgrep) that want to tie notifications
to specific commit links and that could not be done
with the on-the-fly rebase.

Signed-off-by: michael sorens <msorens@chef.io>