automate-cs-oc-erchef: fix GEM_PATH in run hook (#4473)
This was using pkgPathFor on core/oc_erchef which doesn't exist. The
package is actually called chef/oc_erchef.

Astute readers may wonder whether we even need to step this variable
if all the tests pare passing even though it was previously
incorrectly set. That issue is left as an exercise for the reader.

Signed-off-by: Steven Danna <steve@chef.io>