Encrypt secrets that are put in the secrets store (#4495)
This modifies the secrets store/secrets helper utilities to encrypt secrets that are stored on disk. Any existing secret data will not be encrypted unless it is modified. This is fine, as all existing data is generated and uses the secret helpers to share that data between services.

The reason we are doing this is be able to store customer provided secrets in non plaintext. We will take the secrets that are provided in the config and place them in the secret store encrypted.

This is not a user facing change.
This does not provide any sort of security, as the key is stored on next to the secrets on disk.