Fix segfault during migration of secrets
Users who have secrets configured prior to last release (20210116...)
are stuck in a crash loop for deployment-service. The issue is that the
secrets store is being initialized too late in the process. This change
initializes it sooner. If a nil secrets store is provided, the code now
ignores it and moves on.

Signed-off-by: Jay Mundrawala <jmundrawala@chef.io>