Adjust Buildkite for merge-base
This lets the differential scan truly scan only the files
in the current PR, as compared to the prior approach of just
using the head of master, which would effectively include all
the files since the branch split off from master.

Also relocate and encapsulate commands into script to be more tidy.

Signed-off-by: michael sorens <msorens@chef.io>