Add Integration Testing for the Notifications Service (#4674)
Create a full integration test suite for the notifications service in go. Everything is tested via external interface. This enables reimplementing the notifications-service while maintaining API stability.

No work is done to run this test suite in Ci yet. It will make sense to do so once we've created the reimplementation of the notifications-service.

This suite will probably need some adaption to work against the new implementation of the notifications-service. That will be done when there is a new implementation to adapt to.

* Appease shellcheck
* Add golang integration tests for notifications service
* Organize unit tests
* Add integration tests for notifications service
* More test cases for notifications integration tests