Notifications Rewrite: Service Stub (#4712)
We are rewriting the notifications service in go to align better with our engineering capabilities.

This change creates a `notifications-service2` component in the source tree. It can be compiled and packaged manually, but is hidden from pipeline builds using the `dev_only` option in the habitat `plan.sh` (the `automate-ui-devproxy` uses the same mechanism).

The `notifications-service2` is missing most of the functionality that it will eventually have/need. Currently it only implements configuration file loading. When run as a service it prints its config to the logs in a loop. This is a first step towards implementing the infrastructural components of the new notifications service.

* Habitat boilerplate for notifications-service2
* Add config to notifications2
* Cleanup