Notifications Rewrite: Run Lint/Unit in Ci (#4747)
Add linting and a basic Makefile to notifications-service2. Run the ci make target in Ci, which will run the linter and unit tests both. These are quite fast so it would be slower to parallelize them than to run in serial.

In the process it was necessary to upgrade golangci-lint. A bug in the tool prevented it from analyzing the notifications-service2. I attempted to be as conservative as possible. The (latest version of golangci-lint is 1.37.0)[https://github.com/golangci/golangci-lint/tags]; in this pull request I only upgraded us from 1.21.0 to 1.26.0. The reason for this is each version adds new checks, many of which fail in many components of the codebase. I am trying to keep this patch scoped to only adding testing and static analysis to the new notifications service implementation. In this spirit, I have disabled the new checks that I have found to fail on existing code. Some of these new checks seem kinda dubious, but govet's copylocks is probably one we would like to enable. Nevertheless, this should be done in a separate work stream.

* Add linting to new notifications svc
* Upgrade golangci-lint a little
* Run notifications-service2 lint/unit in Ci
* Update golangci-lint in studiorc
* don't require go fmt -s
* don't require formatting imports
* fix typo
* Make curl available via go_component_make
* gofmt
* disable rowserrcheck linter, it has too many false positives
* Cache newer golangci lint
* Fix lint errors in infra proxy svc
* fix lint errors for lib
* Fix lint failures for compliance service
* fix node manager lint errors
* fix ingest service lint errors
* Don't return err on invalid search (?)